Added optional queue system for input.
* Player.cs:
  Initial version.

* SpaceTapper.csproj, Difficulty.cs, GameDifficulty.cs, Difficulty.cs,
  DifficultySettings.cs, GameDifficulty.cs, DifficultySettings.cs:
  

* GameState.cs:
  Converted variables to use _ prefix.
  Start of player implementation.

* Game.cs:
  Added EndFrame event.
  Converted variables to use _ prefix.

* Input.cs:
  Added optional input queue.
  Documentation.

* MenuState.cs, DifficultyState.cs:
  Added handler for escape.
  Converted variables to use _ prefix.

* State.cs, ButtonList.cs:
  Converted variables to use _ prefix.

* Button.cs:
  Added rounding for the text origin, to avoid blurry fonts.

+++ SpaceTapper/Source/UI/Button.cs
@@ -76,7 +76,7 @@ namespace SpaceTapper.UI
 		public void Center()
 		{
 			var b = Text.GetLocalBounds();
-			Text.Origin = new Vector2f(b.Width / 2, b.Height / 2);
+			Text.Origin = new Vector2f((int)b.Width / 2, (int)b.Height / 2);
 		}
 
 		protected override void UpdateSelf(double dt)
